clarify "1500" hours was before writing paper
note that after experimentation was done, MPI revoked my access to the MPI infra even though I still was working on THEIR PROJECT UNPAID. They even revoked my institutional email and I even lost access to the publication email for prepartion of presentation. So, this 1500 hours was when we finished experiemntation and started writing paper. The process after that took about 2 month (still unpaid -- I was paid 0 cents for the entire internship)